+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # Changelog
+v2.2.0
+----------------
+- Sowmewhat limited support for Deflate compression.
+- API for parsing ZIP entries.
+
 v2.1.0
 ----------------
 - ZIP containers support.